almost all moves implemented (except for castling)
i had one != instead of == in the coordinate enumerator in the move utils and it messed up my moves. anyways, it seems to be working now. i also had an infinite loop due to result-gamestates checking whether they are game-over which requires checking the legal moves which requires checking whether they put the king in check which is a thing of the future gamestate. maybe i can skip the gamestate part and just check the check-map. anyways, i'm super tired now